# src/main/resources/v1/eric-oss-auto-config-consistency-openapi.yaml
...
openapi: "3.0.2"
servers:
  - url: http://eacc/v1
info:
  x-api-id: 8608e72f-121d-4049-be33-7eb2d3cf5594
  x-audience: external-public
  description: |
    Ericsson Automated Configuration Consistency (EACC) provides self-service analytics 
    to enable users to uncover baseline consistency errors. EACC supports the scheduling 
    of CM consistency audits which detect inconsistencies based on user-supplied rules. 
    These inconsistencies can be corrected manually by the user. 
    Any changes can also be manually rolled back by the user.
  version: 0.1.5
  title: Ericsson Automated Configuration Consistency
  license:
    name: COPYRIGHT Ericsson 2024
    url: https://www.ericsson.com/en/legal
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: pdlonteamq@pdl.internal.ericsson.com
...
paths:
...
  /v1/executions:
    get:
      tags:
        - executions
      summary: Get all the executions or executions for the specified Job name.
      description: |
        Get all the configuration consistency executions or executions for the specified Job 
        with a summary of the results. An empty list is returned if no executions exist.
      operationId: getExecutions
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - in: query
          name: jobName
          schema:
            type: string
          description: Name of the Job
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Execution'
              example: [
                {
                  "id": "1",
                  "jobName": "job_market1",
                  "executionType": "Open Loop",
                  "executionStartedAt": "2024-02-19T14:38:00Z",
                  "executionEndedAt": "2024-02-19T14:38:05Z",
                  "consistencyAuditStartedAt": "2024-02-19T14:38:00Z",
                  "consistencyAuditEndedAt": "2024-02-19T14:38:05Z",
                  "executionStatus": "Changes Successful",
                  "totalNodesAudited": 6,
                  "totalNodesFailed": 0,
                  "totalMosAudited": 46,
                  "totalAttributesAudited": 86,
                  "inconsistenciesIdentified": 15
                } ]
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 400,
                "detail": "Invalid resource identifier in url: Only lower case alphanumeric, underscores and dashes allowed, with between 4 -100 characters."
              }
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 404,
                "detail": "Invalid jobName: jobName does not exist"
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get execution(s).",
                "status": 500,
                "detail": "Database operation failed."
              }
  /v1/executions/{id}/audit-results:
    get:
      tags:
        - executions
      summary: Get the configuration consistency audit result.
      description: |
        Get the configuration consistency audit results for the specified execution ID, with optional pagination
        parameters and support for filtering of the data.
      operationId: getAuditResults
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/executionId'
        - in: query
          name: page
          required: false
          schema:
            type: integer
          description: Page Number, defaults to first page if not set
        - in: query
          name: pageSize
          required: false
          schema:
            type: integer
          description: Page Size, defaults to all data if not set.
        - in: query
          name: filter
          required: false
          schema:
            type: string
          description: |
            Filter to apply. The filter format is '\<filterKey\>:\<filterValue\>'
            or if the property should match one in a list of values then the format is 
            '\<filterKey\>:(\<filterValue1\>,\<filterValue2\>,...)'.
            If filtering on multiple filterKeys then the format is
            '\<filter1Key\>:\<filter1Value\>$\<filter2Key\>:\<filter2Value\>'.
            The filter string should be appropriately URL encoded.
            Note the filters are delimited by a $ symbol.
            The filterKey can be one of auditStatus, changeStatus or managedObjectFdn.
            If filtering on the managedObjectFdn then support is also provided for
            partial matching using the % character in the filterValue. If % is 
            included at the start of the filterValue then the filter matches any
            managedObjectFdn ending with the value. If % is included at the end 
            of the filterValue then the filter matches any managedObjectFdn starting 
            with the value. If % is included at both ends then the filter matches any
            managedObjectFdn containing the filterValue. The search for list of managedObjectFdn
            is not supported.
          examples:
            matchingLikeValue:
              value: filter=managedObjectFdn:%EUtranCellFDD%
            matchingLeftLikeValue:
              value: filter=managedObjectFdn:%EUtranCellFDD
            matchingRightLikeValue:
              value: filter=managedObjectFdn:EUtranCellFDD%
            matchSingleValue:
              value: filter=auditStatus:Inconsistent
            matchListOfValues:
              value: filter=managedObjectType(NRCellDU,EUtranCellFDD)
            matchMultipleFilterKeys:
              value: filter=auditStatus:Inconsistent$managedObjectFdn:SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00001,ManagedElement=NR03gNodeBRadio00001,GNBDUFunction=1,NRCellDU=NR03gNodeBRadio00001-1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditResults'
              example: {
                "totalElements": 15,
                "totalPages": 8,
                "currentPage": 1,
                "perPage": 2,
                "hasNext": true,
                "hasPrev": false,
                "results": [
                  {
                    "id": "1",
                    "managedObjectFdn": "SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00001,ManagedElement=NR03gNodeBRadio00001,GNBDUFunction=1,NRCellDU=NR03gNodeBRadio00001-1",
                    "managedObjectType": "NRCellDU",
                    "attributeName": "ailgDlPrbLoadLevel",
                    "currentValue": "0",
                    "preferredValue": "1",
                    "auditStatus": "Consistent",
                    "executionId": "1",
                    "ruleId": "5",
                    "changeStatus": ""
                  },
                  {
                    "id": "2",
                    "managedObjectFdn": "SubNetwork=Europe,SubNetwork=Ireland,MeContext=NR03gNodeBRadio00001,ManagedElement=NR03gNodeBRadio00001,GNBDUFunction=1,NRCellDU=NR03gNodeBRadio00001-2",
                    "managedObjectType": "NRCellDU",
                    "attributeName": "ailgDlPrbLoadLevel",
                    "currentValue": "1",
                    "preferredValue": "1",
                    "auditStatus": "Inconsistent",
                    "executionId": "1",
                    "ruleId": "5",
                    "changeStatus": "Implementation complete"
                  } ]
              }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 400,
                "detail": "Execution ID can only contain numeric characters."
              }
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Execution ID does not exist.",
                "status": 404,
                "detail": "Provide a valid Execution ID."
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get audit results.",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - executions
      summary: Apply or revert changes in the configuration consistency audit result.
      description: |
        Apply changes to fix inconsistencies reported in the audit result or revert changes that 
        were previously applied. One or more changes can be specified by giving a list of audit 
        result IDs or alternatively all inconsistencies can be specified.
      operationId: applyChange
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
        - $ref: '#/components/parameters/executionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovedAuditResults'
            examples:
              selectedChanges:
                value:
                  auditResultIds: [ "1", "2", "3" ]
                  operation: APPLY_CHANGE
              selectedChangesReversion:
                value:
                  auditResultIds: [ "1", "2", "3" ]
                  approveForAll: false
                  operation: REVERT_CHANGE
              allChangesReversion:
                value:
                  auditResultIds: [ ]
                  approveForAll: true
                  operation: REVERT_CHANGE
      responses:
        "202":
          description: Accepted
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Execution ID does not exist.",
                "status": 404,
                "detail": "Provide a valid Execution ID."
              }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 400,
                "detail": "Execution ID can only contain numeric characters."
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to apply change.",
                "status": 500,
                "detail": "Database operation failed."
              }
components:
...
  parameters:
...
    executionId:
      name: id
      in: path
      description: ID of the execution
      required: true
      schema:
        type: string
...
    acceptInHeader:
      name: Accept
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json
    contentTypeInHeader:
      name: Content-Type
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json

  # Common error responses
...
  schemas:
...
    Execution:
      type: object
      required: [ id, jobName, executionType, executionStartedAt, executionEndedAt, consistencyAuditStartedAt, consistencyAuditEndedAt, executionStatus, totalNodesAudited, totalNodesFailed, totalMosAudited, totalAttributesAudited, inconsistenciesIdentified ]
      properties:
        id:
          type: string
          description: The execution ID.
        jobName:
          type: string
          description: The job name for this execution.
        executionType:
          type: string
          description: |
            Type of execution.
        executionStartedAt:
          type: string
          format: date-time
          description: The execution start time.
        executionEndedAt:
          type: string
          format: date-time
          description: The execution end time.
        consistencyAuditStartedAt:
          type: string
          format: date-time
          description: The audit start time.
        consistencyAuditEndedAt:
          type: string
          format: date-time
          description: The audit end time.
        executionStatus:
          type: string
          description: |
            Shows the status of the execution. One of "Audit in Progress", "Audit Successful", "Audit Partially Successful", 
            "Audit Failed", "Changes in Progress", "Changes Failed", "Changes Successful", "Changes Partially Successful",
            "Reversion in Progress", "Reversion Failed", "Reversion Successful" or "Reversion Partially Successful".
        totalNodesAudited:
          type: integer
          description: The total number of nodes audited.
        totalNodesFailed:
          type: integer
          description: The number of nodes that the audit could not be completed for.
        totalMosAudited:
          type: integer
          description: The total number of Managed Objects audited.
        totalAttributesAudited:
          type: integer
          description: The total number of attributes audited.
        inconsistenciesIdentified:
          type: integer
          description: The total number of inconsistencies found in the audit.
    PaginatedAuditResults:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
          description: The total number of audit results for the execution.
        totalPages:
          type: integer
          description: The total number of pages of audit results for the execution.
        currentPage:
          type: integer
          description: The current page of the audit results.
        perPage:
          type: integer
          description: The number of audit results in the page.
        hasNext:
          type: boolean
          description: True if the are more pages, otherwise false.
        hasPrev:
          type: boolean
          description: True if the are previous pages, otherwise false.
        results:
          type: array
          items:
            $ref: '#/components/schemas/AuditResult'
    AuditResult:
      type: object
      required: [ id, managedObjectFdn, managedObjectType, attributeName, currentValue, preferredValue, auditStatus, executionId, ruleId, changeStatus ]
      properties:
        id:
          type: string
          description: The audit ID. This is only unique within the execution.
        managedObjectFdn:
          type: string
          description: The FDN of the Managed Object of the individual audit result.
        managedObjectType:
          type: string
          description: The Managed Object type of the individual audit result.
        attributeName:
          type: string
          description: The attribute name of the individual audit result.
        currentValue:
          type: string
          description: The current value of the attribute.
        preferredValue:
          type: string
          description: The value that the attribute should, have based on the rule.
        auditStatus:
          type: string
          description: The status of the individual audit and can be either "Consistent" or "Inconsistent".
        executionId:
          type: string
          description: The execution ID that this individual audit result is related to.
        ruleId:
          type: string
          description: |
            The rule ID which is used for this individual audit result. The rule ID itself is the ordinal number of
            the rule in the CSV file provided in the Ruleset.
        changeStatus:
          type: string
          description: |
            The status of the change or revert. One of "Implementation in progress", "Implementation complete",
            "Implementation failed", "Reversion in progress", "Reversion complete" or "Reversion failed".
            If no change is applied then this will contain an empty string.
    ApprovedAuditResults:
      type: object
      required: [ auditResultIds ]
      properties:
        auditResultIds:
          type: array
          items:
            type: string
          description: A list of the individual audit result IDs.
        approveForAll:
          type: boolean
          default: false
          description: |
            If set to true then all individual audit results in the execution are changed or reverted. Otherwise
            the list of audit result IDs in the auditResultIds array are used.
        operation:
          $ref: '#/components/schemas/Operation'
    Operation:
      type: string
      description: The operation to apply on the audit result.
      default: APPLY_CHANGE
      enum:
        - APPLY_CHANGE
        - REVERT_CHANGE
    ErrorDetails:
      description: Provides additional information in an error response.
      type: object
      required: [ title, status, detail ]
      properties:
        title:
          $ref: '#/components/schemas/ErrorTitle'
        status:
          $ref: '#/components/schemas/ErrorStatus'
        detail:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorTitle:
      type: string
      description: Short, human-readable summary of the problem.
    ErrorStatus:
      type: integer
      description: HTTP status code.
    ErrorDetail:
      description: |
        Optional human-readable explanation of the problem. if present, ought to focus on helping the client 
        correct the problem, rather than giving debugging information.